% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrection.R
\name{.hermite3_logit}
\alias{.hermite3_logit}
\title{Applying the probabilists' Hermite polynomials (K = 3) to the
transformed estimates of phi}
\usage{
.hermite3_logit(phi)
}
\arguments{
\item{phi}{A numeric value representing the estimate of phi.}
}
\value{
A numeric vector of length 4 containing the Hermite polynomials.
}
\description{
Applying the probabilists' Hermite polynomials (K = 3) to the
transformed estimates of phi
}
\examples{
.hermite3_logit(0.5)

}
\keyword{internal}
